refactor: Did the suggested changes in the PR
