+++ src/App.js
@@ -1,9 +1,10 @@
 import "./App.css";
 import { Routes, Route, Link } from "react-router-dom";
-import Navbar from "./components/Navbar";
-import Signin from "./components/Signin";
-import HomePage from "./pages/HomePage";
-import Signup from "./components/Signup";
+// import Navbar from "./components/Navbar";
+// import Signin from "./components/Signin";
+import { HomePage } from "./pages/HomePage";
+// import Signup from "./components/Signup";
+import { Navbar, Signin, Signup } from "../src/index";
 import MockmanEs from "mockman-js";
 
 function App() {
