+++ src/pages/HomePage.js
@@ -1,9 +1,10 @@
-import Sidebar from "../components/Sidebar";
-import Searchbar from "../components/Searchbar";
-import AddNote from "../components/AddNote";
-import ShowNote from "../components/ShowNote";
+// import Sidebar from "../components/Sidebar";
+// import Searchbar from "../components/Searchbar";
+// import AddNote from "../components/AddNote";
+// import ShowNote from "../components/ShowNote";
+import { Sidebar, Searchbar, AddNote, ShowNote } from "../index";
 
-const HomePage = () => {
+export const HomePage = () => {
   return (
     <div className="app-container pt-1">
       <Sidebar />
@@ -13,4 +14,3 @@ const HomePage = () => {
     </div>
   );
 };
-export default HomePage;
